Merged PR 891915: Add Config XML support (which enables default annotations and custom document variant tag)
## Add Config XML support.

This enables default annotations and custom document variant tag, and establishes structure on how to add future features similarly related to whole document.

The config XML file looks like the following:

```
<?xml version="1.0"?>
<configuration>
  <document>
    <variant>swagger</variant>
    <variant>swagger2</variant>
  </document>
  <operation>
    <annotation tag="$default">
      <param name="param1" required="false" cref="T:System.String" in="header">Added Header Param.</param>
    </annotation>
    <annotation tag="Sample Group 1">
      <param name="param2" required="false" cref="T:System.String" in="header">Added Header Param.</param>
    </annotation>
  </operation>
</configuration>
```

### Structure
The `<document>` section contains config that is related to the whole `OpenApiV3Document`. In this case, the document variant tags are specified. This is used to enable the forked document feature (multiple OpenAPI documents from one XML).

The `<operation>` section contains config that will be specified in the `Operation` level. In this case, we add the annotation of `<param name="param1">` to all the operations (specified by `$default` tag) and the annotation of `<param name="param2">` to operations with in `Sample Group 1` (determined by reading the `group` tag in `ApplyGroupsAsTagFilter.cs`).

### Behavior & Extensibility

The common annotations are read in order, so the behavior is always predictable. The order can be important in the case where, for example, `<group>` tag itself is being added to the operations in the common annotations.

During the specification generation, the `<operation>` section is read using `IOperationConfigFilter`. This makes this config file extensible for future additions. For example, we might want to add the feature to, say, remove certain things from (or do certain other things to) all operations. This can simply be added under `<operation>` and then the implementation for that logic will go into another `ApplyXXXXXXConfigFilter : IOperationConfigFilter`.

Related work items: #13256107