Merged PR 968993: Allow custom attributes for each document variant using document config
_Despite containing many files, this change is relatively small. Only a couple of lines are updated in most XMLs to reflect the new way of supplying security/version attributes._

### Allow custom attributes for each document variant using document config

These custom attributes are populated in the `DocumentVariantInfo` project for the customer of the generator to use.

Since the attributes are associated to each variant (as opposed to each API), the attributes should NOT be present in the XML documentation.
```
<swagger>Group1</swagger>
```

The attributes will be populated for each variant as long as they are specified in the config file. For example,

```
<?xml version="1.0"?>
<configuration>
  <document>
    <variant>
      <name>swagger</name>
      <options>
        <option security="sg1" version="V2">Group1</option>
        <option security="sgnotexist" version="V2">Group2</option>
      </options>
    </variant>
    <variant>
      <name>swagger2</name>
      <options>
        <option security="sg1" version="V2">Group1</option>
        <option security="sgnotexist" version="V2">Group2</option>
      </options>
    </variant>
  </document>
  <operation>
  </operation>
</configuration>
```

The attributes can also be put in the main annotation XML (to support backward compat) and we will pick it up. Any conflicts will result in a blank document and a failed path generation result.

```
<swagger security="sg1" version="V2">Group1</swagger>
```

These attributes can technically be anything. Our current use in MSSwagger is for security group and versioning, so this change will allow us to support this existing functionality.

Related work items: #14170235