Add predefined operationId support.
If the operation described in the XML file has an <operationId> tag, it
is used, otherwise (keeping the original behaviour) the id is generated
using the path (<url>) and the operation type (<verb>).

In order to do this, the following changes have been introduced:
+ Add UsePredefinedOperationIdFilter class which creates operations using
<operationId> tag.

+ Add CreateOperationMetaFilter which executes a list of operation creation
filters. The first successful filter is used to generate the operations.
This mechanism is used to support both the predefined and the generated
operation id: first, the predefined operation id based filter is used, and
if it fails, the operation id generation based filter is executed.

+ Add ICreateOperationPreprocessingOperationFilter to support applicable
check in operation create filters in a general way.

+ Extend OperationHandler with GetOperationId function.
+ Slight refactor.
+ Small fix in GetUrl function.
+ Add extensive unit tests

+ Remove duplicated code from BranchOptionalPathParametersFilter and use
OperationHandler existing functionality.